+++ libopendmi/include/opendmi/error.h
@@ -31,6 +31,7 @@ typedef enum dmi_error_code
     DMI_ERROR_FILE_WRITE,            ///< Unable to write file
     DMI_ERROR_FILE_DUP,              ///< Unable to clone file handle
     DMI_ERROR_FILE_MAP,              ///< Unable to map file
+    DMI_ERROR_EPS_NOT_FOUND,         ///< Entry point structure not found
     DMI_ERROR_UNKNOWN_EPS_ANCHOR,    ///< Unknown entry point structure anchor
     DMI_ERROR_INVALID_EPS_LENGTH,    ///< Invalid entry point structure length
     DMI_ERROR_INVALID_EPS_CHECKSUM,  ///< Invalid entry point structure checksum